connected quotes with main, made all structuras in file struct.h
