+++ minishell.h
@@ -6,7 +6,7 @@
 /*   By: ineimatu <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/18 13:00:48 by ineimatu          #+#    #+#             */
-/*   Updated: 2024/07/23 17:17:16 by ineimatu         ###   ########.fr       */
+/*   Updated: 2024/07/24 13:17:21 by ineimatu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,24 +15,11 @@
 
 # include <stdio.h>
 # include <stdlib.h>
+# include <unistd.h>
 # include "readline/readline.h"
 # include "readline/history.h"
 #include "lexer.h"
-
-typedef struct s_envp
-{
-	char 	*key;
-	char	*value;
-	struct s_envp 	*next;
-} t_envp;
-
-typedef	struct s_info
-{
-	char	*rl;
-	t_lex_lst	*tokens;
-	int	ex_stat;
-	t_envp	*envp;
-} t_info;
+#include "struct.h"
 
 // initial functions
 void	start_reading(t_info *info);
