+++ struct.h
@@ -1,21 +1,50 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lexer.h                                            :+:      :+:    :+:   */
+/*   struct.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ineimatu <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/07/23 10:36:08 by ineimatu          #+#    #+#             */
-/*   Updated: 2024/07/24 13:10:19 by ineimatu         ###   ########.fr       */
+/*   Created: 2024/07/24 13:07:01 by ineimatu          #+#    #+#             */
+/*   Updated: 2024/07/24 13:12:39 by ineimatu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef LEXER_H
-# define LEXER_H
+#ifndef STRUCT_H
+# define STRUCT_H
+
 #include "minishell.h"
-#include "struct.h"
 
+typedef struct s_envp
+{
+	char 	*key;
+	char	*value;
+	struct s_envp 	*next;
+} t_envp;
 
-void	valid_line(t_info *info);
+typedef	enum s_token
+{
+	PIPE = 1,
+	OUTPUT,
+	APPEND,
+	INPUT,
+	HEREDOC,
+} t_token;
+
+typedef struct s_lex_lst
+{
+	char	*str;
+	t_token	token;
+	int		elem;
+	struct s_lex_lst	*next;
+}	t_lex_lst;
+
+typedef	struct s_info
+{
+	char	*rl;
+	t_lex_lst	*tokens;
+	int	ex_stat;
+	t_envp	*envp;
+} t_info;
 
 #endif
